requirements.txt
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.4
bcrypt==3.1.7
beautifulsoup4==4.8.2
black==19.3b0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
drf-yasg==1.17.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
isort==5.4.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
libevdev==0.5
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.19.1
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
pdfarranger==1.4.2
pexpect==4.6.0
pikepdf==1.10.3+dfsg
Pillow==7.0.0
pipenv==11.9.0
protobuf==3.6.1
psycopg2==2.8.5
pycairo==1.16.2
pycodestyle==2.6.0
pycups==1.9.73
pyflakes==2.1.1
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-git==2018.2.1
pytz==2019.3
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
smmap==3.0.4
soupsieve==1.9.5
sqlparse==0.3.1
systemd-python==234
toml==0.10.1
ttystatus==0.38
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
usb-creator==0.3.7
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
webencodings==0.5.1
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
